Refactoring classes
